+++ src/ScottPlot5/ScottPlot5/Interfaces/ISignalXYSource.cs
@@ -23,6 +23,11 @@ public interface ISignalXYSource
     public double YScale { get; set; }
 
     /// <summary>
+    /// Multiply X values by this scale factor (before applying offset)
+    /// </summary>
+    public double XScale { get; set; }
+
+    /// <summary>
     /// Do not display data below this index
     /// </summary>
     public int MinimumIndex { get; set; }